Move file left behind by merge conflict to correct new location (#741)
A file was added to cadl-openapi (now called cadl-autorest) due
to a merge conflict between PRs that was resolved without moving
the new file over to the new location. The result is that these
tests don't run now. This moves the file to the correct location,
updates the test name to match the new package name, and reacts
to _delete to delete rename.