Revert accidental change that started running decorators/onBuild in LS (#784)
A mistake during a refactoring caused us to start running decorators and
onBuild in language server. We do intend to do this soon, but doing it
prematurely revealed a version problem where we can't safely run decorators
that load a local Cadl compiler while using the globally installed language
server.

So, for now, revert back to not running decorators, and we will look at how
to deal with the version mismatch as part of doing it again deliberately.