Fix issue where dlls were overwriting each other on build.
`dotnet build -o` is unfortunately not really useful on multi-targeted
libs. Fix is to `dotnet pack -o` as in the ProviderHub.Common project, and
disable GeneratePackageOnBuild likewise too.

This change also adds ExcludeAssets=Runtime to prevent msbuild dlls and
depedencies from being unnecessarily copied to the output directory. This
didn't end up being the source of the problem, but these were getting
consumed by the signing process and that led me down a path to investigate
where the dlls were coming from.